refactor(sdf): move Layer into sdf to mirror C++ SdfLayer
The `Layer` type is functionally `SdfLayer` — single-layer identity, an `AbstractData` backend, the authoring surface,
and persistence — and belongs alongside the other sdf primitives. Moves `Layer`, `LayerFormat`, `AuthoringError`, and
the authoring helpers into `src/sdf/layer.rs`, leaving `src/layer.rs` to host only `Collector` and the dependency-walking
loader (no direct C++ analog).
