+++ src/layer.rs
@@ -3,20 +3,20 @@
 //! Given a root USD file, [`Collector`] uses an [`ar::Resolver`] to recursively
 //! resolve and load every layer the stage depends on — following sublayers,
 //! references, and payloads across files and formats (`.usda`, `.usdc`, `.usd`,
-//! `.usdz`). The result is a [`Vec`] of [`Layer`]s, each wrapping a parsed
-//! [`AbstractData`] with its resolved identity. Cycles are detected and
-//! skipped automatically.
+//! `.usdz`). The result is a [`Vec`] of [`sdf::Layer`]s, each wrapping a
+//! parsed [`sdf::AbstractData`] with its resolved identity. Cycles are
+//! detected and skipped automatically.
+//!
+//! The [`sdf::Layer`] type itself, along with its authoring surface and
+//! persistence (`save`/`save_as`), lives in [`crate::sdf`] — the Rust
+//! equivalent of C++ `SdfLayer`.
 
 use std::collections::{HashMap, HashSet};
 use std::io::Cursor;
 
 use anyhow::{bail, Context, Result};
 
-use crate::ar::{self, Resolver};
-use crate::expr;
-use crate::sdf::schema::{ChildrenKey, FieldKey};
-use crate::sdf::{self, AbstractData, LayerData, Path, Value};
-use crate::{usda, usdc, usdz};
+use crate::{ar, expr, sdf, usda, usdc, usdz};
 
 /// The kind of layer dependency that triggered a composition error.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -61,593 +61,8 @@ pub enum Error {
         /// What kind of composition arc declared this dependency.
         kind: DependencyKind,
         /// The prim that declared this arc (`None` for sublayers).
-        prim_path: Option<Path>,
+        prim_path: Option<sdf::Path>,
     },
-}
-
-/// A single loaded layer in the composition.
-pub struct Layer {
-    /// Resolved, canonical identifier for this layer.
-    pub identifier: String,
-    /// The parsed scene description data. Kept crate-private so external
-    /// callers cannot swap the backing box or write to it directly,
-    /// bypassing the authoring API's bookkeeping invariants
-    /// (`primChildren`, `propertyChildren`, ancestor specifiers, …).
-    pub(crate) data: LayerData,
-}
-
-impl Layer {
-    /// Borrow the underlying [`AbstractData`] backend for read-only inspection.
-    /// Callers that need typed views should prefer [`Layer::prim`] /
-    /// [`Layer::attr`] / etc.
-    pub fn data(&self) -> &dyn AbstractData {
-        self.data.as_ref()
-    }
-}
-
-/// Persistent format for a saved layer.
-///
-/// Used by [`Layer::save_as`] and [`LayerFormat::from_extension`] to make
-/// the writer's format choice explicit. For `.usd` — which the AOUSD Core
-/// Spec permits to be either text (§16.2) or binary (§16.3) — this lets the
-/// caller pick.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum LayerFormat {
-    /// Text format (`.usda`).
-    Usda,
-    /// Binary crate format (`.usdc`).
-    Usdc,
-    /// Archive/package format (`.usdz`).
-    Usdz,
-}
-
-impl LayerFormat {
-    /// Infer the most likely format from a file extension.
-    ///
-    /// `.usda` → [`LayerFormat::Usda`], `.usdc` → [`LayerFormat::Usdc`],
-    /// `.usdz` → [`LayerFormat::Usdz`]. The ambiguous `.usd` extension defaults
-    /// to [`LayerFormat::Usdc`], matching the C++ reference implementation
-    /// (`USD_WRITE_NEW_USD_FILES_AS_BINARY=true`). Unknown or missing
-    /// extensions return `None`.
-    pub fn from_extension(ext: &str) -> Option<Self> {
-        match ext.to_ascii_lowercase().as_str() {
-            "usda" => Some(Self::Usda),
-            "usdc" | "usd" => Some(Self::Usdc),
-            "usdz" => Some(Self::Usdz),
-            _ => None,
-        }
-    }
-}
-
-impl Layer {
-    fn new(identifier: impl Into<String>, data: LayerData) -> Self {
-        Self {
-            identifier: identifier.into(),
-            data,
-        }
-    }
-
-    /// Save the layer to disk, dispatching on the destination's extension.
-    ///
-    /// - `.usda` → text writer
-    /// - `.usdc` → binary crate writer
-    /// - `.usd` → binary crate writer (see below)
-    /// - `.usdz` → archive containing one `.usdc` entry named after the layer's
-    ///   final path component (with the extension swapped to `.usdc`)
-    ///
-    /// # `.usd` format choice
-    ///
-    /// Per the AOUSD Core Specification, `.usd` is valid for **either** text
-    /// (§16.2: "stored with the .usda or .usd extension") **or** binary
-    /// (§16.3: "represented with the .usdc or .usd extension"). The reader
-    /// side auto-detects via magic-byte sniffing.
-    ///
-    /// `save()` defaults to binary for `.usd`, matching Pixar's C++ USD default
-    /// (`USD_WRITE_NEW_USD_FILES_AS_BINARY=true`). To force a specific format
-    /// — for example, to save text to a `.usd` path — use [`Layer::save_as`].
-    pub fn save(&self, path: impl AsRef<std::path::Path>) -> Result<()> {
-        let path = path.as_ref();
-        let ext = path.extension().and_then(|e| e.to_str()).unwrap_or_default();
-        let format = LayerFormat::from_extension(ext).ok_or_else(|| match ext {
-            "" => anyhow::anyhow!("layer path {} has no extension; cannot choose format", path.display()),
-            other => anyhow::anyhow!("unsupported layer extension {other:?} for save (expected usda/usdc/usd/usdz)"),
-        })?;
-        self.save_as(path, format)
-    }
-
-    /// Save the layer to disk using an explicit format.
-    ///
-    /// Unlike [`Layer::save`], the destination path's extension is not
-    /// consulted — the writer strictly uses `format`. This is the path for:
-    ///
-    /// - Writing text to a `.usd` file: `save_as(path, LayerFormat::Usda)`
-    /// - Writing binary to a `.usd` file (explicit, not just default):
-    ///   `save_as(path, LayerFormat::Usdc)`
-    /// - Emitting an atypical extension in general.
-    ///
-    /// Note that the reader's magic-byte auto-detection will still read the
-    /// result correctly regardless of the filename extension.
-    pub fn save_as(&self, path: impl AsRef<std::path::Path>, format: LayerFormat) -> Result<()> {
-        let path = path.as_ref();
-        match format {
-            LayerFormat::Usda => usda::TextWriter::write_to_file(self.data.as_ref(), path),
-            LayerFormat::Usdc => usdc::CrateWriter::write_to_file(self.data.as_ref(), path),
-            LayerFormat::Usdz => {
-                let stem = path.file_stem().and_then(|s| s.to_str()).unwrap_or("layer");
-                let mut buf = Vec::new();
-                usdc::CrateWriter::write(self.data.as_ref(), &mut Cursor::new(&mut buf))?;
-                let mut archive = crate::usdz::ArchiveWriter::create(path)?;
-                archive.add_layer(&format!("{stem}.usdc"), &buf)?;
-                archive.finish()?;
-                Ok(())
-            }
-        }
-    }
-}
-
-/// Errors raised by [`Layer`]'s authoring methods.
-#[derive(Debug, thiserror::Error)]
-#[non_exhaustive]
-pub enum AuthoringError {
-    /// The layer's backend is not an in-memory [`sdf::Data`] store. Layers
-    /// loaded from files via [`Collector`] are read-only.
-    #[error("layer {identifier} is read-only; authoring is not supported")]
-    ReadOnly {
-        /// The layer's resolved identifier.
-        identifier: String,
-    },
-
-    /// The given path is not valid for the requested authoring operation.
-    /// Prim authoring requires an absolute, non-root, non-property path;
-    /// property authoring requires a property path.
-    #[error("path {path} is not valid here: {reason}")]
-    InvalidPath {
-        /// The offending path.
-        path: Path,
-        /// A short, human-readable reason describing what was expected.
-        reason: &'static str,
-    },
-}
-
-/// Authoring methods. These require the layer's backend to be a writable
-/// in-memory [`sdf::Data`] store (created via [`Layer::new_anonymous`]).
-/// File-loaded layers return [`AuthoringError::ReadOnly`].
-///
-/// Mirrors `SdfLayer`'s authoring surface: [`Layer::create_prim`],
-/// [`Layer::override_prim`], [`Layer::create_attr`], and
-/// [`Layer::create_relationship`] parallel `SdfCreatePrimInLayer` /
-/// `SdfAttributeSpec::New` / `SdfRelationshipSpec::New`. `primChildren`
-/// and `propertyChildren` ordering is maintained automatically: ancestor
-/// prim specs are auto-created as `over` when missing.
-///
-/// # Reading file-loaded layers
-///
-/// The typed-view lookups ([`Layer::prim`], [`Layer::attr`],
-/// [`Layer::relationship`], [`Layer::pseudo_root`]) currently borrow an
-/// underlying [`sdf::Data`] directly. Layers loaded by `usda::TextReader` or
-/// `usdc::CrateData` keep their own internal spec storage and do not yet
-/// expose an `sdf::Data` view, so the lookup methods return `None` on
-/// file-loaded layers. Use [`crate::stage::Stage`]'s path-keyed query API
-/// for inspection in the meantime.
-//
-// TODO: materialize `TextReader` / `CrateData` into `sdf::Data` (or expose
-// one via `AbstractData::as_data`) so the typed views work uniformly across
-// in-memory and file-loaded layers, matching C++ `SdfLayer::GetPrimAtPath`.
-impl Layer {
-    /// Create a blank in-memory writable layer with the given identifier.
-    ///
-    /// The layer's pseudo-root spec is pre-populated so layer-level metadata
-    /// (`defaultPrim`, `subLayers`, time codes, …) can be authored via
-    /// [`Layer::pseudo_root_mut`] immediately.
-    pub fn new_anonymous(identifier: impl Into<String>) -> Self {
-        let mut data = sdf::Data::new();
-        data.create_spec(Path::abs_root(), sdf::SpecType::PseudoRoot);
-        Self::new(identifier, Box::new(data))
-    }
-
-    /// Create or upgrade a prim spec at `path` with the given specifier and
-    /// `typeName`. Any missing ancestor specs are created as `over` and the
-    /// parent's `primChildren` is updated.
-    ///
-    /// Mirrors C++ `SdfCreatePrimInLayer` + `SetSpecifier` + `SetTypeName`.
-    pub fn create_prim(
-        &mut self,
-        path: impl Into<Path>,
-        specifier: sdf::Specifier,
-        type_name: impl Into<String>,
-    ) -> Result<sdf::PrimSpecMut<'_>, AuthoringError> {
-        let path: Path = path.into();
-        let type_name: String = type_name.into();
-
-        require_prim_path(&path)?;
-        let data = self.writable_data_mut()?;
-        ensure_prim_chain(data, &path)?;
-
-        let spec = data.spec_mut(&path).expect("just ensured");
-        spec.add(FieldKey::Specifier, Value::Specifier(specifier));
-        spec.add(FieldKey::TypeName, Value::Token(type_name));
-
-        Ok(spec.as_prim_mut().expect("type guaranteed by ensure_prim_chain"))
-    }
-
-    /// Ensure a prim spec exists at `path` with specifier `over`. Existing
-    /// `def` / `class` specs are left as-is; this only creates missing ones.
-    /// Ancestor specs are auto-created as `over` as well.
-    ///
-    /// Mirrors C++ `UsdStage::OverridePrim` semantics at the layer tier.
-    pub fn override_prim(&mut self, path: impl Into<Path>) -> Result<sdf::PrimSpecMut<'_>, AuthoringError> {
-        let path: Path = path.into();
-
-        require_prim_path(&path)?;
-        let data = self.writable_data_mut()?;
-        ensure_prim_chain(data, &path)?;
-
-        let spec = data.spec_mut(&path).expect("just ensured");
-        Ok(spec.as_prim_mut().expect("type guaranteed by ensure_prim_chain"))
-    }
-
-    /// Create an attribute spec at `path` (a property path like
-    /// `/World/Mesh.points`). The owning prim is auto-created as `over` if
-    /// missing, and the prim's `propertyChildren` is updated.
-    ///
-    /// `typeName` and `variability` are construction parameters (matching
-    /// C++ `SdfAttributeSpec::New`); there is no post-hoc setter for either.
-    pub fn create_attr(
-        &mut self,
-        path: impl Into<Path>,
-        type_name: impl Into<String>,
-        variability: sdf::Variability,
-        custom: bool,
-    ) -> Result<sdf::AttributeSpecMut<'_>, AuthoringError> {
-        let path: Path = path.into();
-        let type_name: String = type_name.into();
-
-        let (prim_path, property_name) = split_property_path(&path)?;
-        let data = self.writable_data_mut()?;
-        require_spec_type_or_absent(data, &path, sdf::SpecType::Attribute)?;
-        validate_token_vec(data, &prim_path, ChildrenKey::PropertyChildren)?;
-        ensure_prim_chain(data, &prim_path)?;
-
-        add_to_token_vec(
-            data.spec_mut(&prim_path).expect("ensure_prim_chain created it"),
-            &prim_path,
-            ChildrenKey::PropertyChildren,
-            &property_name,
-        )?;
-
-        if !data.has_spec(&path) {
-            data.create_spec(path.clone(), sdf::SpecType::Attribute);
-        }
-        let spec = data.spec_mut(&path).expect("just ensured");
-        spec.add(FieldKey::TypeName, Value::Token(type_name));
-        if variability != sdf::Variability::Varying {
-            spec.add(FieldKey::Variability, Value::Variability(variability));
-        } else {
-            spec.remove(FieldKey::Variability.as_str());
-        }
-        if custom {
-            spec.add(FieldKey::Custom, Value::Bool(true));
-        } else {
-            spec.remove(FieldKey::Custom.as_str());
-        }
-
-        Ok(spec
-            .as_attr_mut()
-            .expect("type guaranteed by require_spec_type_or_absent"))
-    }
-
-    /// Create a relationship spec at `path`. The owning prim is auto-created
-    /// as `over` if missing, and the prim's `propertyChildren` is updated.
-    ///
-    /// `variability` and `custom` are construction parameters, matching C++
-    /// `SdfCreateRelationshipInLayer`.
-    pub fn create_relationship(
-        &mut self,
-        path: impl Into<Path>,
-        variability: sdf::Variability,
-        custom: bool,
-    ) -> Result<sdf::RelationshipSpecMut<'_>, AuthoringError> {
-        let path: Path = path.into();
-
-        let (prim_path, property_name) = split_property_path(&path)?;
-        let data = self.writable_data_mut()?;
-        require_spec_type_or_absent(data, &path, sdf::SpecType::Relationship)?;
-        validate_token_vec(data, &prim_path, ChildrenKey::PropertyChildren)?;
-        ensure_prim_chain(data, &prim_path)?;
-
-        add_to_token_vec(
-            data.spec_mut(&prim_path).expect("ensure_prim_chain created it"),
-            &prim_path,
-            ChildrenKey::PropertyChildren,
-            &property_name,
-        )?;
-
-        if !data.has_spec(&path) {
-            data.create_spec(path.clone(), sdf::SpecType::Relationship);
-        }
-        let spec = data.spec_mut(&path).expect("just ensured");
-        if variability != sdf::Variability::Varying {
-            spec.add(FieldKey::Variability, Value::Variability(variability));
-        } else {
-            spec.remove(FieldKey::Variability.as_str());
-        }
-        if custom {
-            spec.add(FieldKey::Custom, Value::Bool(true));
-        } else {
-            spec.remove(FieldKey::Custom.as_str());
-        }
-
-        Ok(spec
-            .as_relationship_mut()
-            .expect("type guaranteed by require_spec_type_or_absent"))
-    }
-
-    /// Look up a prim spec at `path`. Returns `None` if no spec exists, the
-    /// spec is not a prim, or the backend is not in-memory writable.
-    pub fn prim(&self, path: impl Into<Path>) -> Option<sdf::PrimSpec<'_>> {
-        let path: Path = path.into();
-        self.data.as_data()?.spec(&path)?.as_prim()
-    }
-
-    /// Mutably look up a prim spec at `path`.
-    pub fn prim_mut(&mut self, path: impl Into<Path>) -> Option<sdf::PrimSpecMut<'_>> {
-        let path: Path = path.into();
-        self.data.as_data_mut()?.spec_mut(&path)?.as_prim_mut()
-    }
-
-    /// Look up an attribute spec at a property path.
-    pub fn attr(&self, path: impl Into<Path>) -> Option<sdf::AttributeSpec<'_>> {
-        let path: Path = path.into();
-        self.data.as_data()?.spec(&path)?.as_attr()
-    }
-
-    /// Mutably look up an attribute spec at a property path.
-    pub fn attr_mut(&mut self, path: impl Into<Path>) -> Option<sdf::AttributeSpecMut<'_>> {
-        let path: Path = path.into();
-        self.data.as_data_mut()?.spec_mut(&path)?.as_attr_mut()
-    }
-
-    /// Look up a relationship spec at a property path.
-    pub fn relationship(&self, path: impl Into<Path>) -> Option<sdf::RelationshipSpec<'_>> {
-        let path: Path = path.into();
-        self.data.as_data()?.spec(&path)?.as_relationship()
-    }
-
-    /// Mutably look up a relationship spec at a property path.
-    pub fn relationship_mut(&mut self, path: impl Into<Path>) -> Option<sdf::RelationshipSpecMut<'_>> {
-        let path: Path = path.into();
-        self.data.as_data_mut()?.spec_mut(&path)?.as_relationship_mut()
-    }
-
-    /// View this layer's root pseudo-spec, which carries layer-wide metadata
-    /// (`defaultPrim`, `subLayers`, time codes, …).
-    pub fn pseudo_root(&self) -> Option<sdf::PseudoRootSpec<'_>> {
-        self.data.as_data()?.spec(&Path::abs_root())?.as_pseudo_root()
-    }
-
-    /// Mutably view this layer's root pseudo-spec. The spec is created on
-    /// first access if missing. Returns [`AuthoringError::ReadOnly`] for
-    /// file-loaded layers.
-    pub fn pseudo_root_mut(&mut self) -> Result<sdf::PseudoRootSpecMut<'_>, AuthoringError> {
-        let data = self.writable_data_mut()?;
-        let root = Path::abs_root();
-        match data.spec_type(&root) {
-            Some(sdf::SpecType::PseudoRoot) => {}
-            Some(_) => {
-                return Err(AuthoringError::InvalidPath {
-                    path: root,
-                    reason: "root spec exists with non-PseudoRoot SpecType",
-                })
-            }
-            None => {
-                data.create_spec(root.clone(), sdf::SpecType::PseudoRoot);
-            }
-        }
-        Ok(data
-            .spec_mut(&root)
-            .expect("just ensured")
-            .as_pseudo_root_mut()
-            .expect("type guaranteed above"))
-    }
-
-    fn writable_data_mut(&mut self) -> Result<&mut sdf::Data, AuthoringError> {
-        let identifier = self.identifier.clone();
-        self.data.as_data_mut().ok_or(AuthoringError::ReadOnly { identifier })
-    }
-}
-
-/// Ensure prim specs exist for every ancestor of `target` and for `target`
-/// itself, creating missing ones as `over`. Updates each parent's
-/// `primChildren` to include the next name along the chain. Idempotent.
-/// `target` must be an absolute, non-root, non-property prim path;
-/// callers should validate via [`require_prim_path`] first.
-///
-/// Errors with [`AuthoringError::InvalidPath`] if any ancestor or `target`
-/// path already holds a spec of a non-prim type — stamping `primChildren`
-/// onto an Attribute or Relationship spec would corrupt the layer.
-fn ensure_prim_chain(data: &mut sdf::Data, target: &Path) -> Result<(), AuthoringError> {
-    let mut chain: Vec<Path> = Vec::new();
-    let mut cursor: Path = target.clone();
-    while !cursor.is_abs_root() {
-        chain.push(cursor.clone());
-        cursor = cursor.parent().expect("non-root has a parent (validated upstream)");
-    }
-    chain.reverse(); // root → leaf
-
-    for (i, child) in chain.iter().enumerate() {
-        let parent_path = if i == 0 { Path::abs_root() } else { chain[i - 1].clone() };
-        let parent_ty = if parent_path.is_abs_root() {
-            sdf::SpecType::PseudoRoot
-        } else {
-            sdf::SpecType::Prim
-        };
-
-        if let Some(existing) = data.spec_type(&parent_path) {
-            if existing != parent_ty {
-                return Err(AuthoringError::InvalidPath {
-                    path: parent_path,
-                    reason: "ancestor spec exists with non-prim SpecType",
-                });
-            }
-        }
-        validate_token_vec(data, &parent_path, ChildrenKey::PrimChildren)?;
-
-        if let Some(existing) = data.spec_type(child) {
-            if existing != sdf::SpecType::Prim {
-                return Err(AuthoringError::InvalidPath {
-                    path: child.clone(),
-                    reason: "spec exists with non-prim SpecType",
-                });
-            }
-        }
-    }
-
-    for (i, child) in chain.iter().enumerate() {
-        let parent_path = if i == 0 { Path::abs_root() } else { chain[i - 1].clone() };
-        let child_name = child.name().expect("non-root has a name").to_owned();
-
-        let parent_ty = if parent_path.is_abs_root() {
-            sdf::SpecType::PseudoRoot
-        } else {
-            sdf::SpecType::Prim
-        };
-        match data.spec_type(&parent_path) {
-            Some(_) => {}
-            None => {
-                data.create_spec(parent_path.clone(), parent_ty);
-            }
-        }
-
-        let parent_spec = data.spec_mut(&parent_path).expect("just ensured");
-        add_to_token_vec(parent_spec, &parent_path, ChildrenKey::PrimChildren, &child_name)?;
-
-        match data.spec_type(child) {
-            Some(_) => {}
-            None => {
-                let spec = data.create_spec(child.clone(), sdf::SpecType::Prim);
-                spec.add(FieldKey::Specifier, Value::Specifier(sdf::Specifier::Over));
-            }
-        }
-    }
-    Ok(())
-}
-
-/// Insert `name` into the `TokenVec` field at `key` on `spec`, creating the
-/// field if absent. No-op if the name is already present.
-fn add_to_token_vec(
-    spec: &mut sdf::Spec,
-    owner_path: &Path,
-    key: ChildrenKey,
-    name: &str,
-) -> Result<(), AuthoringError> {
-    match spec.get_mut(key.as_str()) {
-        Some(Value::TokenVec(v)) => {
-            if !v.iter().any(|n| n == name) {
-                v.push(name.to_owned());
-            }
-        }
-        Some(_) => {
-            return Err(AuthoringError::InvalidPath {
-                path: owner_path.clone(),
-                reason: "child-list field exists with non-TokenVec value",
-            });
-        }
-        None => {
-            spec.add(key, Value::TokenVec(vec![name.to_owned()]));
-        }
-    }
-    Ok(())
-}
-
-/// Verify that an authored child-list field is either absent or a `TokenVec`.
-fn validate_token_vec(data: &sdf::Data, path: &Path, key: ChildrenKey) -> Result<(), AuthoringError> {
-    match data.spec(path).and_then(|spec| spec.get(key.as_str())) {
-        Some(Value::TokenVec(_)) | None => Ok(()),
-        Some(_) => Err(AuthoringError::InvalidPath {
-            path: path.clone(),
-            reason: "child-list field exists with non-TokenVec value",
-        }),
-    }
-}
-
-/// Verify that `path` either holds no spec or holds one of type `expected`.
-fn require_spec_type_or_absent(data: &sdf::Data, path: &Path, expected: sdf::SpecType) -> Result<(), AuthoringError> {
-    match data.spec_type(path) {
-        Some(existing) if existing != expected => Err(AuthoringError::InvalidPath {
-            path: path.clone(),
-            reason: "spec exists with the wrong SpecType",
-        }),
-        _ => Ok(()),
-    }
-}
-
-/// Validate that `path` is an absolute, non-root, non-property path suitable
-/// for prim authoring — every `/`-separated component must be a USD identifier
-/// (no brackets, variant-selection segments, or stray dots).
-fn require_prim_path(path: &Path) -> Result<(), AuthoringError> {
-    if !path.is_abs() || path.is_abs_root() {
-        return Err(AuthoringError::InvalidPath {
-            path: path.clone(),
-            reason: "expected absolute non-root prim path",
-        });
-    }
-    if path.is_property_path() {
-        return Err(AuthoringError::InvalidPath {
-            path: path.clone(),
-            reason: "expected prim path, got property path",
-        });
-    }
-    for segment in path.as_str()[1..].split('/') {
-        if !Path::is_valid_identifier(segment) {
-            return Err(AuthoringError::InvalidPath {
-                path: path.clone(),
-                reason: "prim path component is not a USD identifier",
-            });
-        }
-    }
-    Ok(())
-}
-
-/// Split a property path like `/World/Mesh.points` into `(/World/Mesh,
-/// "points")`. Returns an error if `path` is not an absolute property path
-/// whose owning prim portion is itself a valid prim path.
-fn split_property_path(path: &Path) -> Result<(Path, String), AuthoringError> {
-    if !path.is_property_path() {
-        return Err(AuthoringError::InvalidPath {
-            path: path.clone(),
-            reason: "expected property path",
-        });
-    }
-    let prim_path = path.prim_path();
-    // Owning prim must be an absolute, non-root, non-property path — guards
-    // against relative roots ("A.foo"), root-level properties ("/.foo"), and
-    // paths whose `prim_path()` returned a structurally invalid string.
-    require_prim_path(&prim_path)?;
-    let suffix = path
-        .as_str()
-        .strip_prefix(prim_path.as_str())
-        .and_then(|t| t.strip_prefix('.'))
-        .ok_or(AuthoringError::InvalidPath {
-            path: path.clone(),
-            reason: "malformed property path",
-        })?;
-    // Property names are colon-separated identifiers — reject target/connection
-    // brackets, embedded dots, and other syntax that would round-trip as garbage.
-    if suffix.is_empty() || !suffix.split(':').all(Path::is_valid_identifier) {
-        return Err(AuthoringError::InvalidPath {
-            path: path.clone(),
-            reason: "property name must be a colon-separated identifier",
-        });
-    }
-    Ok((prim_path, suffix.to_owned()))
-}
-
-impl std::fmt::Debug for Layer {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Layer")
-            .field("identifier", &self.identifier)
-            .finish_non_exhaustive()
-    }
 }
 
 /// A dependency discovered while walking a layer's scene graph.
@@ -657,17 +72,17 @@ struct Dependency {
     /// What kind of composition arc declared this dependency.
     kind: DependencyKind,
     /// The prim that declared this arc (`None` for sublayers).
-    prim_path: Option<Path>,
+    prim_path: Option<sdf::Path>,
 }
 
 struct CollectionContext<'a> {
     load_payloads: bool,
-    include_prim_dependency: Option<&'a dyn Fn(&Path) -> bool>,
+    include_prim_dependency: Option<&'a dyn Fn(&sdf::Path) -> bool>,
     on_error: &'a dyn Fn(Error) -> Result<()>,
 }
 
 impl<'a> CollectionContext<'a> {
-    fn with_filter(&self, include_prim_dependency: Option<&'a dyn Fn(&Path) -> bool>) -> Self {
+    fn with_filter(&self, include_prim_dependency: Option<&'a dyn Fn(&sdf::Path) -> bool>) -> Self {
         Self {
             load_payloads: self.load_payloads,
             include_prim_dependency,
@@ -683,14 +98,14 @@ fn strict_error(e: Error) -> Result<()> {
 }
 
 /// Builder for collecting a root layer and all layers it depends on.
-pub struct Collector<'a, R: Resolver, E: Fn(Error) -> Result<()> = StrictErrorHandler> {
+pub struct Collector<'a, R: ar::Resolver, E: Fn(Error) -> Result<()> = StrictErrorHandler> {
     resolver: &'a R,
     load_payloads: bool,
-    include_prim_dependency: Option<&'a dyn Fn(&Path) -> bool>,
+    include_prim_dependency: Option<&'a dyn Fn(&sdf::Path) -> bool>,
     on_error: E,
 }
 
-impl<'a, R: Resolver> Collector<'a, R, StrictErrorHandler> {
+impl<'a, R: ar::Resolver> Collector<'a, R, StrictErrorHandler> {
     /// Creates a collector that errors on any unresolvable dependency and
     /// follows payload arcs. Use [`Collector::on_error`] or
     /// [`Collector::load_payloads`] to override.
@@ -704,7 +119,7 @@ impl<'a, R: Resolver> Collector<'a, R, StrictErrorHandler> {
     }
 }
 
-impl<'a, R: Resolver, E: Fn(Error) -> Result<()>> Collector<'a, R, E> {
+impl<'a, R: ar::Resolver, E: Fn(Error) -> Result<()>> Collector<'a, R, E> {
     /// Controls whether payload dependencies are followed.
     ///
     /// When `load_payloads` is `false`, payload arcs are left as authored
@@ -731,15 +146,15 @@ impl<'a, R: Resolver, E: Fn(Error) -> Result<()>> Collector<'a, R, E> {
     /// `include`. Used by `StageBuilder` to honor `StagePopulationMask` during
     /// root-layer-stack collection so culled-subtree dependencies aren't
     /// resolved or loaded. Crate-internal pending a public design.
-    pub(crate) fn prim_dependency_filter(mut self, include: &'a dyn Fn(&Path) -> bool) -> Self {
+    pub(crate) fn prim_dependency_filter(mut self, include: &'a dyn Fn(&sdf::Path) -> bool) -> Self {
         self.include_prim_dependency = Some(include);
         self
     }
 
     /// Opens a root layer and recursively collects all referenced layers.
     ///
-    /// Returns a [`Vec<Layer>`] with the root (strongest) layer first.
-    pub fn collect(&self, root_path: &str) -> Result<Vec<Layer>> {
+    /// Returns a [`Vec<sdf::Layer>`] with the root (strongest) layer first.
+    pub fn collect(&self, root_path: &str) -> Result<Vec<sdf::Layer>> {
         let mut layers = Vec::new();
         let mut visited = HashSet::new();
         let context = CollectionContext {
@@ -759,11 +174,11 @@ impl<'a, R: Resolver, E: Fn(Error) -> Result<()>> Collector<'a, R, E> {
 
 /// Recursive layer collector.
 fn collect_recursive(
-    resolver: &impl Resolver,
+    resolver: &impl ar::Resolver,
     asset_path: &str,
     anchor: Option<&ar::ResolvedPath>,
     context: &CollectionContext<'_>,
-    layers: &mut Vec<Layer>,
+    layers: &mut Vec<sdf::Layer>,
     visited: &mut HashSet<String>,
 ) -> Result<()> {
     // Create an anchored identifier so relative paths resolve correctly.
@@ -840,21 +255,21 @@ fn collect_recursive(
         )?;
     }
 
-    layers.push(Layer::new(identifier, data));
+    layers.push(sdf::Layer::new(identifier, data));
 
     Ok(())
 }
 
 /// Collects typed dependencies from sublayers, references, and optionally
 /// payloads in a layer.
-fn collect_dependencies(data: &dyn AbstractData, load_payloads: bool) -> Result<Vec<Dependency>> {
+fn collect_dependencies(data: &dyn sdf::AbstractData, load_payloads: bool) -> Result<Vec<Dependency>> {
     let mut deps = Vec::new();
 
-    let root = Path::abs_root();
+    let root = sdf::Path::abs_root();
 
     // Sublayers (layer-level).
-    if let Some(value) = data.try_get(&root, FieldKey::SubLayers.as_str())? {
-        if let Value::StringVec(sub_paths) = value.into_owned() {
+    if let Some(value) = data.try_get(&root, sdf::FieldKey::SubLayers.as_str())? {
+        if let sdf::Value::StringVec(sub_paths) = value.into_owned() {
             for asset_path in sub_paths {
                 deps.push(Dependency {
                     asset_path,
@@ -869,8 +284,8 @@ fn collect_dependencies(data: &dyn AbstractData, load_payloads: bool) -> Result<
     let prim_paths = collect_prim_paths(data)?;
     for prim_path in &prim_paths {
         // References.
-        if let Some(value) = data.try_get(prim_path, FieldKey::References.as_str())? {
-            if let Value::ReferenceListOp(list_op) = value.as_ref() {
+        if let Some(value) = data.try_get(prim_path, sdf::FieldKey::References.as_str())? {
+            if let sdf::Value::ReferenceListOp(list_op) = value.as_ref() {
                 for r in list_op.iter().filter(|r| !r.asset_path.is_empty()) {
                     deps.push(Dependency {
                         asset_path: r.asset_path.clone(),
@@ -883,16 +298,16 @@ fn collect_dependencies(data: &dyn AbstractData, load_payloads: bool) -> Result<
 
         if load_payloads {
             // Payloads.
-            if let Some(value) = data.try_get(prim_path, FieldKey::Payload.as_str())? {
+            if let Some(value) = data.try_get(prim_path, sdf::FieldKey::Payload.as_str())? {
                 match value.as_ref() {
-                    Value::Payload(p) if !p.asset_path.is_empty() => {
+                    sdf::Value::Payload(p) if !p.asset_path.is_empty() => {
                         deps.push(Dependency {
                             asset_path: p.asset_path.clone(),
                             kind: DependencyKind::Payload,
                             prim_path: Some(prim_path.clone()),
                         });
                     }
-                    Value::PayloadListOp(list_op) => {
+                    sdf::Value::PayloadListOp(list_op) => {
                         for p in list_op.iter().filter(|p| !p.asset_path.is_empty()) {
                             deps.push(Dependency {
                                 asset_path: p.asset_path.clone(),
@@ -912,9 +327,9 @@ fn collect_dependencies(data: &dyn AbstractData, load_payloads: bool) -> Result<
 
 /// Collects all prim and variant spec paths by walking `primChildren`,
 /// `variantSetChildren`, and `variantChildren` hierarchies.
-fn collect_prim_paths(data: &dyn AbstractData) -> Result<Vec<Path>> {
+fn collect_prim_paths(data: &dyn sdf::AbstractData) -> Result<Vec<sdf::Path>> {
     let mut result = Vec::new();
-    let mut queue = vec![Path::abs_root()];
+    let mut queue = vec![sdf::Path::abs_root()];
 
     while let Some(path) = queue.pop() {
         if !data.has_spec(&path) {
@@ -922,13 +337,13 @@ fn collect_prim_paths(data: &dyn AbstractData) -> Result<Vec<Path>> {
         }
 
         // Skip the pseudo-root itself but process its children.
-        if path != Path::abs_root() {
+        if path != sdf::Path::abs_root() {
             result.push(path.clone());
         }
 
         // Regular prim children.
-        if let Some(value) = data.try_get(&path, ChildrenKey::PrimChildren.as_str())? {
-            if let Value::TokenVec(children) = value.into_owned() {
+        if let Some(value) = data.try_get(&path, sdf::ChildrenKey::PrimChildren.as_str())? {
+            if let sdf::Value::TokenVec(children) = value.into_owned() {
                 for name in children.iter().rev() {
                     if let Ok(child) = path.append_path(name.as_str()) {
                         queue.push(child);
@@ -938,13 +353,13 @@ fn collect_prim_paths(data: &dyn AbstractData) -> Result<Vec<Path>> {
         }
 
         // Variant set children (e.g. /Prim -> /Prim{setName=}).
-        if let Some(value) = data.try_get(&path, ChildrenKey::VariantSetChildren.as_str())? {
-            if let Value::TokenVec(set_names) = value.into_owned() {
+        if let Some(value) = data.try_get(&path, sdf::ChildrenKey::VariantSetChildren.as_str())? {
+            if let sdf::Value::TokenVec(set_names) = value.into_owned() {
                 for set_name in &set_names {
                     // Variant children within each set (e.g. /Prim{setName=selA}).
                     let set_path = path.append_variant_selection(set_name, "");
-                    if let Some(value) = data.try_get(&set_path, ChildrenKey::VariantChildren.as_str())? {
-                        if let Value::TokenVec(variant_names) = value.into_owned() {
+                    if let Some(value) = data.try_get(&set_path, sdf::ChildrenKey::VariantChildren.as_str())? {
+                        if let sdf::Value::TokenVec(variant_names) = value.into_owned() {
                             for variant_name in &variant_names {
                                 let variant_path = path.append_variant_selection(set_name, variant_name);
                                 queue.push(variant_path);
@@ -963,8 +378,8 @@ fn collect_prim_paths(data: &dyn AbstractData) -> Result<Vec<Path>> {
 ///
 /// Supports `.usda` (text), `.usdc` (binary), `.usd` (auto-detected via magic
 /// bytes), and `.usdz` (archive — reads the first layer). Returns the parsed
-/// data as a boxed [`AbstractData`].
-pub fn open_layer(resolver: &impl Resolver, resolved: &ar::ResolvedPath) -> Result<LayerData> {
+/// data as a boxed [`sdf::AbstractData`].
+pub fn open_layer(resolver: &impl ar::Resolver, resolved: &ar::ResolvedPath) -> Result<sdf::LayerData> {
     let ext = resolved.extension().and_then(|e| e.to_str()).unwrap_or_default();
 
     let mut asset = resolver.open_asset(resolved)?;
@@ -992,10 +407,10 @@ pub fn open_layer(resolver: &impl Resolver, resolved: &ar::ResolvedPath) -> Resu
 }
 
 /// Reads `expressionVariables` from the layer's pseudo-root, if present.
-fn read_expression_variables(data: &dyn AbstractData) -> Result<HashMap<String, Value>> {
-    let root = Path::abs_root();
-    if let Some(value) = data.try_get(&root, FieldKey::ExpressionVariables.as_str())? {
-        if let Value::Dictionary(dict) = value.into_owned() {
+fn read_expression_variables(data: &dyn sdf::AbstractData) -> Result<HashMap<String, sdf::Value>> {
+    let root = sdf::Path::abs_root();
+    if let Some(value) = data.try_get(&root, sdf::FieldKey::ExpressionVariables.as_str())? {
+        if let sdf::Value::Dictionary(dict) = value.into_owned() {
             return Ok(dict);
         }
     }
@@ -1003,14 +418,14 @@ fn read_expression_variables(data: &dyn AbstractData) -> Result<HashMap<String, 
 }
 
 /// Evaluates an expression-valued asset path, or passes it through unchanged.
-fn resolve_expression(path: &str, vars: &HashMap<String, Value>) -> Result<String> {
+fn resolve_expression(path: &str, vars: &HashMap<String, sdf::Value>) -> Result<String> {
     if expr::is_expression(path) {
         let expression = expr::Expr::parse(path).with_context(|| format!("failed to parse expression: {path}"))?;
         let result = expression
             .eval(vars)
             .with_context(|| format!("failed to evaluate expression: {path}"))?;
         match result {
-            Value::String(s) => Ok(s),
+            sdf::Value::String(s) => Ok(s),
             other => bail!("expression must evaluate to a string, got: {other:?}"),
         }
     } else {
@@ -1023,7 +438,8 @@ mod tests {
     use std::cell::RefCell;
 
     use super::*;
-    use crate::ar::DefaultResolver;
+    use crate::ar::{DefaultResolver, Resolver};
+    use crate::sdf::AbstractData;
 
     const VENDOR_COMPOSITION: &str = "vendor/usd-wg-assets/test_assets/foundation/stage_composition";
 
@@ -1336,7 +752,7 @@ mod tests {
         let path = composition_path("references/reference_invalid.usda");
         let resolver = DefaultResolver::new();
         let errors = RefCell::new(0);
-        let include = |path: &Path| path.as_str() == "/World/cube";
+        let include = |path: &sdf::Path| path.as_str() == "/World/cube";
         let layers = Collector::new(&resolver)
             .prim_dependency_filter(&include)
             .on_error(|_| {
@@ -1353,18 +769,18 @@ mod tests {
 
     #[test]
     fn save_dispatches_on_extension() -> Result<()> {
-        use crate::sdf::{self, SpecType, Value};
+        use crate::sdf::{self, SpecType};
 
         let mut data = sdf::Data::new();
         let root = sdf::Path::abs_root();
         let ps = data.create_spec(root, SpecType::PseudoRoot);
-        ps.add("primChildren", Value::TokenVec(vec!["Foo".into()]));
+        ps.add("primChildren", sdf::Value::TokenVec(vec!["Foo".into()]));
         let foo = sdf::path("/Foo")?;
         let sp = data.create_spec(foo, SpecType::Prim);
-        sp.add("specifier", Value::Specifier(sdf::Specifier::Def));
-        sp.add("typeName", Value::Token("Xform".into()));
+        sp.add("specifier", sdf::Value::Specifier(sdf::Specifier::Def));
+        sp.add("typeName", sdf::Value::Token("Xform".into()));
 
-        let layer = Layer::new("test://layer", Box::new(data));
+        let layer = sdf::Layer::new("test://layer", Box::new(data));
         let dir = tempfile::tempdir()?;
 
         let usda_path = dir.path().join("layer-save.usda");
@@ -1388,18 +804,18 @@ mod tests {
 
     #[test]
     fn save_as_usd_writes_binary_and_roundtrips() -> Result<()> {
-        use crate::sdf::{self, SpecType, Value};
+        use crate::sdf::{self, SpecType};
 
         let mut data = sdf::Data::new();
         let root = sdf::Path::abs_root();
         let ps = data.create_spec(root, SpecType::PseudoRoot);
-        ps.add("primChildren", Value::TokenVec(vec!["Bar".into()]));
+        ps.add("primChildren", sdf::Value::TokenVec(vec!["Bar".into()]));
         let bar = sdf::path("/Bar")?;
         let sp = data.create_spec(bar.clone(), SpecType::Prim);
-        sp.add("specifier", Value::Specifier(sdf::Specifier::Def));
-        sp.add("typeName", Value::Token("Cube".into()));
+        sp.add("specifier", sdf::Value::Specifier(sdf::Specifier::Def));
+        sp.add("typeName", sdf::Value::Token("Cube".into()));
 
-        let layer = Layer::new("test://layer-usd", Box::new(data));
+        let layer = sdf::Layer::new("test://layer-usd", Box::new(data));
         let dir = tempfile::tempdir()?;
         let path = dir.path().join("layer-save.usd");
         layer.save(&path)?;
@@ -1419,7 +835,7 @@ mod tests {
         assert_eq!(round.spec_type(&bar), Some(SpecType::Prim));
         assert_eq!(
             round.get(&bar, "typeName").unwrap().into_owned(),
-            Value::Token("Cube".into())
+            sdf::Value::Token("Cube".into())
         );
         Ok(())
     }
@@ -1429,7 +845,7 @@ mod tests {
         use crate::sdf::{self, SpecType};
         let mut data = sdf::Data::new();
         data.create_spec(sdf::Path::abs_root(), SpecType::PseudoRoot);
-        let layer = Layer::new("test://layer", Box::new(data));
+        let layer = sdf::Layer::new("test://layer", Box::new(data));
         let err = layer.save("/tmp/openusd-bad.xyz").unwrap_err();
         assert!(err.to_string().contains("unsupported"));
     }
@@ -1439,21 +855,21 @@ mod tests {
     /// auto-detects it as text via the absence of the USDC magic.
     #[test]
     fn save_as_forces_text_to_usd_extension() -> Result<()> {
-        use crate::sdf::{self, SpecType, Value};
+        use crate::sdf::{self, SpecType};
 
         let mut data = sdf::Data::new();
         let root = sdf::Path::abs_root();
         let ps = data.create_spec(root, SpecType::PseudoRoot);
-        ps.add("primChildren", Value::TokenVec(vec!["Text".into()]));
+        ps.add("primChildren", sdf::Value::TokenVec(vec!["Text".into()]));
         let prim = sdf::path("/Text")?;
         let sp = data.create_spec(prim.clone(), SpecType::Prim);
-        sp.add("specifier", Value::Specifier(sdf::Specifier::Def));
-        sp.add("typeName", Value::Token("Xform".into()));
+        sp.add("specifier", sdf::Value::Specifier(sdf::Specifier::Def));
+        sp.add("typeName", sdf::Value::Token("Xform".into()));
 
-        let layer = Layer::new("test://text-as-usd", Box::new(data));
+        let layer = sdf::Layer::new("test://text-as-usd", Box::new(data));
         let dir = tempfile::tempdir()?;
         let path = dir.path().join("text-as-usd.usd");
-        layer.save_as(&path, LayerFormat::Usda)?;
+        layer.save_as(&path, sdf::LayerFormat::Usda)?;
 
         // Emitted bytes must NOT start with the binary magic — they're text.
         let bytes = std::fs::read(&path)?;
@@ -1470,20 +886,20 @@ mod tests {
         assert_eq!(round.spec_type(&prim), Some(SpecType::Prim));
         assert_eq!(
             round.get(&prim, "typeName").unwrap().into_owned(),
-            Value::Token("Xform".into())
+            sdf::Value::Token("Xform".into())
         );
         Ok(())
     }
 
     #[test]
     fn layer_format_from_extension_matches_spec() {
-        assert_eq!(LayerFormat::from_extension("usda"), Some(LayerFormat::Usda));
-        assert_eq!(LayerFormat::from_extension("usdc"), Some(LayerFormat::Usdc));
-        assert_eq!(LayerFormat::from_extension("usd"), Some(LayerFormat::Usdc));
-        assert_eq!(LayerFormat::from_extension("USDA"), Some(LayerFormat::Usda));
-        assert_eq!(LayerFormat::from_extension("usdz"), Some(LayerFormat::Usdz));
-        assert_eq!(LayerFormat::from_extension("xyz"), None);
-        assert_eq!(LayerFormat::from_extension(""), None);
+        assert_eq!(sdf::LayerFormat::from_extension("usda"), Some(sdf::LayerFormat::Usda));
+        assert_eq!(sdf::LayerFormat::from_extension("usdc"), Some(sdf::LayerFormat::Usdc));
+        assert_eq!(sdf::LayerFormat::from_extension("usd"), Some(sdf::LayerFormat::Usdc));
+        assert_eq!(sdf::LayerFormat::from_extension("USDA"), Some(sdf::LayerFormat::Usda));
+        assert_eq!(sdf::LayerFormat::from_extension("usdz"), Some(sdf::LayerFormat::Usdz));
+        assert_eq!(sdf::LayerFormat::from_extension("xyz"), None);
+        assert_eq!(sdf::LayerFormat::from_extension(""), None);
     }
 
     // -----------------------------------------------------------------------
@@ -1492,7 +908,7 @@ mod tests {
 
     #[test]
     fn create_prim_basic() -> Result<()> {
-        let mut layer = Layer::new_anonymous("anon.usda");
+        let mut layer = sdf::Layer::new_anonymous("anon.usda");
         let mut prim = layer.create_prim("/World", sdf::Specifier::Def, "Xform")?;
         prim.set_kind("group");
 
@@ -1505,7 +921,7 @@ mod tests {
 
     #[test]
     fn auto_ancestor_chain() -> Result<()> {
-        let mut layer = Layer::new_anonymous("anon.usda");
+        let mut layer = sdf::Layer::new_anonymous("anon.usda");
         layer.create_prim("/A/B/C", sdf::Specifier::Def, "Mesh")?;
 
         // Leaf is Def; ancestors are Over.
@@ -1523,7 +939,7 @@ mod tests {
 
     #[test]
     fn prim_children() -> Result<()> {
-        let mut layer = Layer::new_anonymous("anon.usda");
+        let mut layer = sdf::Layer::new_anonymous("anon.usda");
         layer.create_prim("/World", sdf::Specifier::Def, "Xform")?;
         layer.create_prim("/World/Mesh", sdf::Specifier::Def, "Mesh")?;
         layer.create_prim("/World/Cube", sdf::Specifier::Def, "Cube")?;
@@ -1541,7 +957,7 @@ mod tests {
 
     #[test]
     fn property_children() -> Result<()> {
-        let mut layer = Layer::new_anonymous("anon.usda");
+        let mut layer = sdf::Layer::new_anonymous("anon.usda");
         layer.create_prim("/Mesh", sdf::Specifier::Def, "Mesh")?;
         layer.create_attr("/Mesh.points", "point3f[]", sdf::Variability::Varying, false)?;
         layer.create_attr("/Mesh.normals", "normal3f[]", sdf::Variability::Varying, false)?;
@@ -1564,7 +980,7 @@ mod tests {
 
     #[test]
     fn relationship_variability() -> Result<()> {
-        let mut layer = Layer::new_anonymous("anon.usda");
+        let mut layer = sdf::Layer::new_anonymous("anon.usda");
         layer.create_relationship("/Mesh.material:binding", sdf::Variability::Uniform, false)?;
 
         let rel = layer.relationship("/Mesh.material:binding").expect("relationship");
@@ -1573,39 +989,39 @@ mod tests {
         layer.create_relationship("/Mesh.material:binding", sdf::Variability::Varying, false)?;
         let rel = layer.relationship("/Mesh.material:binding").expect("relationship");
         assert_eq!(rel.variability(), sdf::Variability::Varying);
-        assert!(rel.get(FieldKey::Variability.as_str()).is_none());
+        assert!(rel.get(sdf::FieldKey::Variability.as_str()).is_none());
         Ok(())
     }
 
     #[test]
     fn bad_prim_children_errors() {
-        let mut layer = Layer::new_anonymous("anon.usda");
+        let mut layer = sdf::Layer::new_anonymous("anon.usda");
         layer
             .data
             .as_data_mut()
             .unwrap()
-            .spec_mut(&Path::abs_root())
+            .spec_mut(&sdf::Path::abs_root())
             .unwrap()
-            .add(ChildrenKey::PrimChildren, Value::String("bad".into()));
+            .add(sdf::ChildrenKey::PrimChildren, sdf::Value::String("bad".into()));
 
         let err = layer.create_prim("/A", sdf::Specifier::Def, "Xform").unwrap_err();
-        assert!(matches!(err, AuthoringError::InvalidPath { .. }));
+        assert!(matches!(err, sdf::AuthoringError::InvalidPath { .. }));
 
         let root = layer
             .data
             .as_data()
             .unwrap()
-            .spec(&Path::abs_root())
+            .spec(&sdf::Path::abs_root())
             .expect("pseudo-root present");
         assert!(matches!(
-            root.get(ChildrenKey::PrimChildren.as_str()),
-            Some(Value::String(value)) if value == "bad"
+            root.get(sdf::ChildrenKey::PrimChildren.as_str()),
+            Some(sdf::Value::String(value)) if value == "bad"
         ));
     }
 
     #[test]
     fn bad_property_children_errors() -> Result<()> {
-        let mut layer = Layer::new_anonymous("anon.usda");
+        let mut layer = sdf::Layer::new_anonymous("anon.usda");
         layer.create_prim("/Mesh", sdf::Specifier::Def, "Mesh")?;
         layer
             .data
@@ -1613,38 +1029,38 @@ mod tests {
             .unwrap()
             .spec_mut(&sdf::path("/Mesh").unwrap())
             .unwrap()
-            .add(ChildrenKey::PropertyChildren, Value::String("bad".into()));
+            .add(sdf::ChildrenKey::PropertyChildren, sdf::Value::String("bad".into()));
 
         let err = layer
             .create_relationship("/Mesh.material:binding", sdf::Variability::Varying, false)
             .unwrap_err();
-        assert!(matches!(err, AuthoringError::InvalidPath { .. }));
+        assert!(matches!(err, sdf::AuthoringError::InvalidPath { .. }));
 
         let data = layer.data.as_data().unwrap();
         assert!(data.spec(&sdf::path("/Mesh.material:binding").unwrap()).is_none());
         let mesh = data.spec(&sdf::path("/Mesh").unwrap()).expect("prim present");
         assert!(matches!(
-            mesh.get(ChildrenKey::PropertyChildren.as_str()),
-            Some(Value::String(value)) if value == "bad"
+            mesh.get(sdf::ChildrenKey::PropertyChildren.as_str()),
+            Some(sdf::Value::String(value)) if value == "bad"
         ));
         Ok(())
     }
 
     #[test]
     fn attr_samples() -> Result<()> {
-        let mut layer = Layer::new_anonymous("anon.usda");
+        let mut layer = sdf::Layer::new_anonymous("anon.usda");
         layer.create_prim("/Sphere", sdf::Specifier::Def, "Sphere")?;
 
         let mut radius = layer.create_attr("/Sphere.radius", "double", sdf::Variability::Varying, false)?;
-        radius.set_default(Value::Double(2.5));
-        radius.set_time_sample(0.0, Value::Double(1.0));
-        radius.set_time_sample(10.0, Value::Double(3.0));
+        radius.set_default(sdf::Value::Double(2.5));
+        radius.set_time_sample(0.0, sdf::Value::Double(1.0));
+        radius.set_time_sample(10.0, sdf::Value::Double(3.0));
         // Out-of-order insert lands in sorted position.
-        radius.set_time_sample(5.0, Value::Double(2.0));
+        radius.set_time_sample(5.0, sdf::Value::Double(2.0));
 
-        let read = layer.attr("/Sphere.radius").expect("attr");
+        let read = layer.attribute("/Sphere.radius").expect("attr");
         assert_eq!(read.type_name(), Some("double"));
-        assert_eq!(read.default(), Some(&Value::Double(2.5)));
+        assert_eq!(read.default(), Some(&sdf::Value::Double(2.5)));
         let samples = read.time_samples().expect("samples authored");
         let times: Vec<f64> = samples.iter().map(|(t, _)| *t).collect();
         assert_eq!(times, vec![0.0, 5.0, 10.0]);
@@ -1653,7 +1069,7 @@ mod tests {
 
     #[test]
     fn wrong_ancestor_type() {
-        let mut layer = Layer::new_anonymous("anon.usda");
+        let mut layer = sdf::Layer::new_anonymous("anon.usda");
         // Plant a non-Prim spec at a prim-shaped path through the public data API.
         layer
             .data
@@ -1662,16 +1078,16 @@ mod tests {
             .create_spec(sdf::path("/A").unwrap(), sdf::SpecType::Attribute);
 
         let err = layer.create_prim("/A/B", sdf::Specifier::Def, "Xform").unwrap_err();
-        assert!(matches!(err, AuthoringError::InvalidPath { .. }));
+        assert!(matches!(err, sdf::AuthoringError::InvalidPath { .. }));
         let err = layer
             .create_attr("/A.x", "double", sdf::Variability::Varying, false)
             .unwrap_err();
-        assert!(matches!(err, AuthoringError::InvalidPath { .. }));
+        assert!(matches!(err, sdf::AuthoringError::InvalidPath { .. }));
     }
 
     #[test]
     fn failed_prim_chain_creation_is_atomic() {
-        let mut layer = Layer::new_anonymous("anon.usda");
+        let mut layer = sdf::Layer::new_anonymous("anon.usda");
         let bad_child = sdf::path("/A/B").unwrap();
         layer
             .data
@@ -1680,32 +1096,40 @@ mod tests {
             .create_spec(bad_child, sdf::SpecType::Attribute);
 
         let err = layer.create_prim("/A/B", sdf::Specifier::Def, "Xform").unwrap_err();
-        assert!(matches!(err, AuthoringError::InvalidPath { .. }));
+        assert!(matches!(err, sdf::AuthoringError::InvalidPath { .. }));
 
         let data = layer.data.as_data().unwrap();
         assert!(data.spec(&sdf::path("/A").unwrap()).is_none());
-        let root = data.spec(&Path::abs_root()).expect("pseudo-root present");
-        assert!(root.get(ChildrenKey::PrimChildren.as_str()).is_none());
+        let root = data.spec(&sdf::Path::abs_root()).expect("pseudo-root present");
+        assert!(root.get(sdf::ChildrenKey::PrimChildren.as_str()).is_none());
     }
 
     #[test]
     fn nan_time_sample() -> Result<()> {
-        let mut layer = Layer::new_anonymous("anon.usda");
+        let mut layer = sdf::Layer::new_anonymous("anon.usda");
         layer.create_prim("/Sphere", sdf::Specifier::Def, "Sphere")?;
         let mut r = layer.create_attr("/Sphere.radius", "double", sdf::Variability::Varying, false)?;
 
-        r.set_time_sample(1.0, Value::Double(1.0));
-        r.set_time_sample(f64::NAN, Value::Double(99.0));
-        r.set_time_sample(2.0, Value::Double(2.0));
+        r.set_time_sample(1.0, sdf::Value::Double(1.0));
+        r.set_time_sample(f64::NAN, sdf::Value::Double(99.0));
+        r.set_time_sample(2.0, sdf::Value::Double(2.0));
         // NaN does not collide with finite samples — both finite values survive.
-        let samples = layer.attr("/Sphere.radius").unwrap().time_samples().unwrap().to_vec();
+        let samples = layer
+            .attribute("/Sphere.radius")
+            .unwrap()
+            .time_samples()
+            .unwrap()
+            .to_vec();
         let finite: Vec<f64> = samples.iter().map(|(t, _)| *t).filter(|t| t.is_finite()).collect();
         assert_eq!(finite, vec![1.0, 2.0]);
 
         // erase_time_sample(NaN) can find the NaN entry via total_cmp.
-        assert!(layer.attr_mut("/Sphere.radius").unwrap().erase_time_sample(f64::NAN));
+        assert!(layer
+            .attribute_mut("/Sphere.radius")
+            .unwrap()
+            .erase_time_sample(f64::NAN));
         let times: Vec<f64> = layer
-            .attr("/Sphere.radius")
+            .attribute("/Sphere.radius")
             .unwrap()
             .time_samples()
             .unwrap()
@@ -1718,7 +1142,7 @@ mod tests {
 
     #[test]
     fn override_prim() -> Result<()> {
-        let mut layer = Layer::new_anonymous("anon.usda");
+        let mut layer = sdf::Layer::new_anonymous("anon.usda");
         layer.override_prim("/A/B")?;
 
         assert_eq!(layer.prim("/A").and_then(|p| p.specifier()), Some(sdf::Specifier::Over));
@@ -1739,7 +1163,7 @@ mod tests {
 
     #[test]
     fn pseudo_root_metadata() -> Result<()> {
-        let mut layer = Layer::new_anonymous("anon.usda");
+        let mut layer = sdf::Layer::new_anonymous("anon.usda");
         {
             let mut root = layer.pseudo_root_mut().expect("writable");
             root.set_default_prim("World");
@@ -1763,72 +1187,72 @@ mod tests {
 
     #[test]
     fn read_only_layer() {
-        // A layer wrapping an empty Data via the public AbstractData trait but
+        // A layer wrapping an empty Data via the public sdf::AbstractData trait but
         // *without* the writable hook would normally come from a file reader.
-        // We simulate by wrapping a custom no-op AbstractData.
+        // We simulate by wrapping a custom no-op sdf::AbstractData.
         struct ReadOnly;
-        impl AbstractData for ReadOnly {
-            fn has_spec(&self, _: &Path) -> bool {
+        impl sdf::AbstractData for ReadOnly {
+            fn has_spec(&self, _: &sdf::Path) -> bool {
                 false
             }
-            fn has_field(&self, _: &Path, _: &str) -> bool {
+            fn has_field(&self, _: &sdf::Path, _: &str) -> bool {
                 false
             }
-            fn spec_type(&self, _: &Path) -> Option<sdf::SpecType> {
+            fn spec_type(&self, _: &sdf::Path) -> Option<sdf::SpecType> {
                 None
             }
-            fn try_get(&self, _: &Path, _: &str) -> Result<Option<std::borrow::Cow<'_, Value>>> {
+            fn try_get(&self, _: &sdf::Path, _: &str) -> Result<Option<std::borrow::Cow<'_, sdf::Value>>> {
                 Ok(None)
             }
-            fn list(&self, _: &Path) -> Option<Vec<String>> {
+            fn list(&self, _: &sdf::Path) -> Option<Vec<String>> {
                 None
             }
-            fn paths(&self) -> Vec<Path> {
+            fn paths(&self) -> Vec<sdf::Path> {
                 Vec::new()
             }
         }
 
-        let mut layer = Layer::new("file.usda", Box::new(ReadOnly));
+        let mut layer = sdf::Layer::new("file.usda", Box::new(ReadOnly));
         let err = layer.create_prim("/World", sdf::Specifier::Def, "Xform").unwrap_err();
-        assert!(matches!(err, AuthoringError::ReadOnly { .. }));
+        assert!(matches!(err, sdf::AuthoringError::ReadOnly { .. }));
     }
 
     #[test]
     fn invalid_paths() {
-        let mut layer = Layer::new_anonymous("anon.usda");
+        let mut layer = sdf::Layer::new_anonymous("anon.usda");
         assert!(matches!(
             layer.create_prim("/", sdf::Specifier::Def, "Xform").unwrap_err(),
-            AuthoringError::InvalidPath { .. }
+            sdf::AuthoringError::InvalidPath { .. }
         ));
         assert!(matches!(
             layer.create_prim("/A.foo", sdf::Specifier::Def, "Xform").unwrap_err(),
-            AuthoringError::InvalidPath { .. }
+            sdf::AuthoringError::InvalidPath { .. }
         ));
         assert!(matches!(
             layer
                 .create_attr("/A", "double", sdf::Variability::Varying, false)
                 .unwrap_err(),
-            AuthoringError::InvalidPath { .. }
+            sdf::AuthoringError::InvalidPath { .. }
         ));
         // Relative property paths must error, not panic in ensure_prim_chain.
         assert!(matches!(
             layer
                 .create_attr("A.foo", "double", sdf::Variability::Varying, false)
                 .unwrap_err(),
-            AuthoringError::InvalidPath { .. }
+            sdf::AuthoringError::InvalidPath { .. }
         ));
         assert!(matches!(
             layer
                 .create_relationship("A.foo", sdf::Variability::Varying, false)
                 .unwrap_err(),
-            AuthoringError::InvalidPath { .. }
+            sdf::AuthoringError::InvalidPath { .. }
         ));
         // Root-level property paths (`/.foo`) must also error, not panic.
         assert!(matches!(
             layer
                 .create_attr("/.foo", "double", sdf::Variability::Varying, false)
                 .unwrap_err(),
-            AuthoringError::InvalidPath { .. }
+            sdf::AuthoringError::InvalidPath { .. }
         ));
         // Target-bracket property paths slip past `is_property_path` because the
         // tail after the last `.` is alphanumeric — split_property_path must reject them.
@@ -1836,26 +1260,26 @@ mod tests {
             layer
                 .create_attr("/A.rel[/Target].attr", "double", sdf::Variability::Varying, false)
                 .unwrap_err(),
-            AuthoringError::InvalidPath { .. }
+            sdf::AuthoringError::InvalidPath { .. }
         ));
         // Variant-selection segments aren't USD identifiers — reject them in prim paths.
         assert!(matches!(
             layer
                 .create_prim("/A{x=y}/B", sdf::Specifier::Def, "Xform")
                 .unwrap_err(),
-            AuthoringError::InvalidPath { .. }
+            sdf::AuthoringError::InvalidPath { .. }
         ));
     }
 
     #[test]
     fn usda_roundtrip() -> Result<()> {
-        let mut layer = Layer::new_anonymous("scene.usda");
+        let mut layer = sdf::Layer::new_anonymous("scene.usda");
         layer.pseudo_root_mut().unwrap().set_default_prim("World");
         layer.create_prim("/World", sdf::Specifier::Def, "Xform")?;
         let mut sphere = layer.create_prim("/World/Sphere", sdf::Specifier::Def, "Sphere")?;
         sphere.set_kind("component");
         let mut radius = layer.create_attr("/World/Sphere.radius", "double", sdf::Variability::Varying, false)?;
-        radius.set_default(Value::Double(1.5));
+        radius.set_default(sdf::Value::Double(1.5));
         let material = sdf::path("/World/Material")?;
         layer.create_prim(&material, sdf::Specifier::Def, "Material")?;
         let mut binding =
@@ -1870,7 +1294,7 @@ mod tests {
         surface.set_connection_paths([sdf::path("/World/Material.outputs:surface")?]);
 
         let tmp = std::env::temp_dir().join("openusd_authoring_roundtrip.usda");
-        layer.save_as(&tmp, LayerFormat::Usda)?;
+        layer.save_as(&tmp, sdf::LayerFormat::Usda)?;
 
         let parsed = usda::TextReader::read(&tmp)?;
         assert_eq!(parsed.spec_type(&sdf::path("/World")?), Some(sdf::SpecType::Prim));
@@ -1884,18 +1308,18 @@ mod tests {
         );
         assert_eq!(
             parsed
-                .get(&sdf::Path::abs_root(), FieldKey::DefaultPrim.as_str())?
+                .get(&sdf::Path::abs_root(), sdf::FieldKey::DefaultPrim.as_str())?
                 .into_owned(),
-            Value::Token("World".into())
+            sdf::Value::Token("World".into())
         );
         match parsed
             .get(
                 &sdf::path("/World/Sphere.material:binding")?,
-                FieldKey::TargetPaths.as_str(),
+                sdf::FieldKey::TargetPaths.as_str(),
             )?
             .into_owned()
         {
-            Value::PathListOp(op) => {
+            sdf::Value::PathListOp(op) => {
                 assert!(op.explicit);
                 assert_eq!(op.explicit_items, vec![material]);
             }
@@ -1904,11 +1328,11 @@ mod tests {
         match parsed
             .get(
                 &sdf::path("/World/Sphere.inputs:surface")?,
-                FieldKey::ConnectionPaths.as_str(),
+                sdf::FieldKey::ConnectionPaths.as_str(),
             )?
             .into_owned()
         {
-            Value::PathListOp(op) => {
+            sdf::Value::PathListOp(op) => {
                 assert!(op.explicit);
                 assert_eq!(op.explicit_items, vec![sdf::path("/World/Material.outputs:surface")?]);
             }
